+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * Fixed plugin initialization in CLI
     * Fixed broken logic in `Page::topParent()` when dealing with first-level pages
     * Fixed broken `Flex Page` authorization for groups
+    * Fixed missing `onAdminSave` and `onAdminAfterSave` events when using `Flex Pages` and `Flex Users` [flex-objects#58](https://github.com/trilbymedia/grav-plugin-flex-objects/issues/58)
 
 # v1.7.0-rc.12
 ## 06/08/2020